# Optimize virtual calls