Testing out reducing the number of virtual calls.

Current master:

|                  Type |          Method |      N | BaseIsPending |      Mean | Code Size |
|---------------------- |---------------- |------- |-------------- |----------:|----------:|
|          AwaitPending | ProtoPromise_V2 | 100000 |             ? | 202.81 ms |  13,241 B |
|          AsyncPending | ProtoPromise_V2 | 100000 |             ? |  96.68 ms |   7,565 B |
|         AsyncResolved | ProtoPromise_V2 | 100000 |             ? |  17.27 ms |   2,215 B |
|         AwaitResolved | ProtoPromise_V2 | 100000 |             ? |  12.13 ms |   2,845 B |
|   ContinueWithPending | ProtoPromise_V2 | 100000 |         False | 270.02 ms |  13,204 B |
| ContinueWithFromValue | ProtoPromise_V2 | 100000 |         False |  23.91 ms |   8,180 B |
|  ContinueWithResolved | ProtoPromise_V2 | 100000 |         False |  30.81 ms |  10,187 B |
|   ContinueWithPending | ProtoPromise_V2 | 100000 |          True | 272.20 ms |  11,523 B |
| ContinueWithFromValue | ProtoPromise_V2 | 100000 |          True | 125.91 ms |      40 B |
|  ContinueWithResolved | ProtoPromise_V2 | 100000 |          True | 152.35 ms |  10,026 B |

This PR:

|                  Type |          Method |      N | BaseIsPending |      Mean | Code Size |
|---------------------- |---------------- |------- |-------------- |----------:|----------:|
|          AwaitPending | ProtoPromise_V2 | 100000 |             ? | 180.02 ms |  11,540 B |
|          AsyncPending | ProtoPromise_V2 | 100000 |             ? | 101.60 ms |   7,853 B |
|         AsyncResolved | ProtoPromise_V2 | 100000 |             ? |  17.83 ms |   2,677 B |
|         AwaitResolved | ProtoPromise_V2 | 100000 |             ? |  10.45 ms |   1,906 B |
|   ContinueWithPending | ProtoPromise_V2 | 100000 |         False | 278.52 ms |  13,354 B |
| ContinueWithFromValue | ProtoPromise_V2 | 100000 |         False |  25.21 ms |   8,357 B |
|  ContinueWithResolved | ProtoPromise_V2 | 100000 |         False |  29.74 ms |  10,538 B |
|   ContinueWithPending | ProtoPromise_V2 | 100000 |          True | 278.36 ms |  11,925 B |
| ContinueWithFromValue | ProtoPromise_V2 | 100000 |          True | 131.80 ms |  10,043 B |
|  ContinueWithResolved | ProtoPromise_V2 | 100000 |          True | 154.04 ms |  10,227 B |