@timcassell 2022-01-16
It's strange, but reducing the virtual calls from 2 to 1 didn't help at all. The only winners here are `AwaitPending` and `AwaitResolved`. I will try to see if I can improve just those and revert the rest.